added prelimenary version of get vertebra orientation
